Added better detection for packages that are unbuildable
This change adds a conditional to check / ensure that required packages
have a safe gaurd / sanity check from setting a value that is unpossible
to be built. This new check will look to see if versions are required to
be less than the greater value or if the less than value is equal to the
greater than value then it will be set as == with no other constraints.