Updated yaprt to do additional return code checking
It is possible to have a returned object become unavailable due to
transient issues upstream. When building the report yaprt does a
head on an object and reviews the returned code. If the code == 200
yaprt assumes that the file will be remotely available. Sadly this
assumption is not true. Before getting the contents of a requirements
file yaprt will need to do another return code check.