added check for repo + branch before processing
The report building process assumed that the user would enter only
valid data within the context of the repo and branch. This
assumption is wrong and yaprt needs to do check to see that the branch
exists as it was inputed prior to moving on and processing. Anything
passed to yaprt should be treated as if it was an absolute.