Updated all of yaprt to clone locally and build using bdist_wheel
as the primary means to create wheels. This change makes yaprt
faster to initial deployment and way faster on updates. This also
extends yaprt to allow it to work strictly using the git protocol
making it no longer subject to github and git.openstack.org.

On this road the commit also removed unused code paths and
does some class and method cleanup. The new bits are still
functional with old reports / data structures.

Major version bump.