Updated to allow for ref spec deployments
Yaprt will now understand and use upstream git repo ref specs from
gerrit. This will allow a deployer / developer to use yaprt to wrap
complext repos and integration requirements with upstream patches
to see how they perform and wether or not they will build in a
greater deployment infrastructure. The changes allow a user to
set the *_git_repo_branch variable to a comma seperated list of
branches to checkout against a ref spec.

if a refspec is used and or multiple branches / commits yaprt
will now create an integration branch and apply the patches to the
source in a single ingration branch and build wheels using only that
now patched branch. This keeps the git history clean while also
allowing for further patching to be done on that source tree as
needed.