Fixed issue when a repo has a "." in the name
Originally yaprt would split on the extension and set the repo
name to the first entry in the split index. This causes issues and
confusion when dealing with repos that have a "." in the name as
the content after the period is assumed to be the "extension".

This change updates the way that the repo name is set and stored
for use later.