Add subdirectory processing for requirement files
This change adds subdirectory processing for requirements within a
given requirement file. Because a requirement file could have an
item like '-e path/package' within it yaprt needs to be updated to
support such a requirement. The change simply allows yaprt to
reprocess a requirement entry as a new subdirectory package.