Perform force-clean on indirect dependencies
Packages in the requirements (build_reqs.txt) could have dependencies on
packages that are not in the requirements. Previously force-clean does
not take into account these indrect dependencies and would not search
and remove existing packages for them.

This fix performs force-clean after pip wheel build and goes through the
list of all built wheels instead of only the requirements when searching
for packages that should be removed.